:art: refactor auth utils
